pyproject.toml
[project]
name = "fastapi-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "authlib>=1.3.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "itsdangerous>=2.1.2",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<4.1.0",
    "python-multipart>=0.0.6",
    "psycopg[binary]>=3.2.9",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.24",
    "testcontainers>=4.12.0",
    "uvicorn>=0.35.0",
]
